Restrict access to SetWarriorMessage in Client
The current warrior, like all state information stored by client,
should
not be directly managed outside of the Client class. The Current
Warrior
should be set by simply sending a SetWarriorMessage to the
server (which the Client class will use to update the WarriorMessage
once it receives the ClientStateUpdateMessage back from the server).